Fix chats migration type mismatch and add missing imports to OrderController

+++ database/migrations/2025_02_19_154259_create_chats_table.php
@@ -11,8 +11,8 @@ class CreateChatsTable extends Migration
         if (!Schema::hasTable('chats')) {
             Schema::create('chats', function (Blueprint $table) {
                 $table->bigIncrements('id');
-                $table->unsignedInteger('sender_id');
-                $table->unsignedInteger('receiver_id');
+                $table->unsignedBigInteger('sender_id');
+                $table->unsignedBigInteger('receiver_id');
                 $table->timestamps();
 
                 $table->foreign('sender_id')
